Moves ParquetStackGrid and ParquetStatusGrid into Parquets namespace.
